+++ package/example/lib/screens/widgets/demo_scaffold.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class DemoScaffold extends StatelessWidget {
+class DemoScaffold extends StatefulWidget {
   final String title;
   final List<Widget> children;
   final Widget? bottom;
@@ -13,6 +13,13 @@ class DemoScaffold extends StatelessWidget {
     this.bottom,
     this.sidePadding = 16.0,
   }) : super(key: key);
+
+  @override
+  State<DemoScaffold> createState() => _DemoScaffoldState();
+}
+
+class _DemoScaffoldState extends State<DemoScaffold> {
+  final scrollController = ScrollController();
 
   @override
   Widget build(BuildContext context) {
@@ -28,8 +35,8 @@ class DemoScaffold extends StatelessWidget {
               children: [
                 const BackButton(),
                 Expanded(
-                  child:
-                      Text(title, style: Theme.of(context).textTheme.headline6),
+                  child: Text(widget.title,
+                      style: Theme.of(context).textTheme.headline6),
                 ),
               ],
             ),
@@ -39,27 +46,29 @@ class DemoScaffold extends StatelessWidget {
       body: Center(
         child: Scrollbar(
           thumbVisibility: true,
+          controller: scrollController,
           child: SingleChildScrollView(
+            controller: scrollController,
             child: Padding(
-              padding: EdgeInsets.symmetric(horizontal: sidePadding) +
+              padding: EdgeInsets.symmetric(horizontal: widget.sidePadding) +
                   const EdgeInsets.symmetric(vertical: 16.0),
               child: Column(
                 mainAxisSize: MainAxisSize.min,
-                children: children,
+                children: widget.children,
               ),
             ),
           ),
         ),
       ),
-      bottomNavigationBar: (bottom != null)
+      bottomNavigationBar: (widget.bottom != null)
           ? Container(
               decoration: const BoxDecoration(
                 border: Border(top: BorderSide(color: Colors.black12)),
               ),
               child: Padding(
                 padding: const EdgeInsets.only(top: 16, bottom: 16) +
-                    EdgeInsets.symmetric(horizontal: sidePadding),
-                child: bottom,
+                    EdgeInsets.symmetric(horizontal: widget.sidePadding),
+                child: widget.bottom,
               ),
             )
           : null,