+++ package/lib/src/breadcrumbs.dart
@@ -27,7 +27,8 @@ class BreadcrumbItem<T> with Diagnosticable {
     super.debugFillProperties(properties);
     properties.add(StringProperty('text', text));
     properties.add(DiagnosticsProperty<T?>('data', data));
-    properties.add(ObjectFlagProperty<ValueChanged<T>?>.has('onSelect', onSelect));
+    properties
+        .add(ObjectFlagProperty<ValueChanged<T>?>.has('onSelect', onSelect));
   }
 }
 
@@ -75,7 +76,8 @@ class Breadcrumbs<T> extends StatelessWidget {
     final effectiveTheme = theme ?? BreadcrumbsThemeData();
     final textStyle = effectiveTheme.getTextStyle(context);
     final activeColor = effectiveTheme.getItemColor(context, textColor);
-    final inactiveColor = effectiveTheme.getInactiveItemColor(context, textColor);
+    final inactiveColor =
+        effectiveTheme.getInactiveItemColor(context, textColor);
     final separatorColor = effectiveTheme.getSeparatorColor(context, textColor);
     final overlayColor = effectiveTheme.getOverlayColor(context, textColor);
 
@@ -104,18 +106,23 @@ class Breadcrumbs<T> extends StatelessWidget {
                   scrollDirection: Axis.horizontal,
                   itemCount: items.length,
                   itemBuilder: (BuildContext context, int index) {
-                    final textColor = (index == (items.length - 1)) ? activeColor : inactiveColor;
+                    final textColor = (index == (items.length - 1))
+                        ? activeColor
+                        : inactiveColor;
 
                     return TextButton(
                       style: ButtonStyle(
                         overlayColor: MaterialStateProperty.all(overlayColor),
                         minimumSize: (effectiveTheme.itemMinimumSize != null)
-                            ? MaterialStateProperty.all(effectiveTheme.itemMinimumSize)
+                            ? MaterialStateProperty.all(
+                                effectiveTheme.itemMinimumSize)
                             : null,
                         padding: (effectiveTheme.itemPadding != null)
-                            ? MaterialStateProperty.all(effectiveTheme.itemPadding)
+                            ? MaterialStateProperty.all(
+                                effectiveTheme.itemPadding)
                             : null,
-                        tapTargetSize: effectiveTheme.getItemTapTargetSize(context),
+                        tapTargetSize:
+                            effectiveTheme.getItemTapTargetSize(context),
                       ),
                       child: Text(
                         items[index].text,
@@ -135,8 +142,10 @@ class Breadcrumbs<T> extends StatelessWidget {
                       size: effectiveTheme.getSeparatorIconSize(context),
                     ),
                   ),
-                  headerBuilder: (_) => SizedBox(width: effectiveTheme.getLeadingSpacing(context)),
-                  footerBuilder: (_) => SizedBox(width: effectiveTheme.getTrailingSpacing(context)),
+                  headerBuilder: (_) => SizedBox(
+                      width: effectiveTheme.getLeadingSpacing(context)),
+                  footerBuilder: (_) => SizedBox(
+                      width: effectiveTheme.getTrailingSpacing(context)),
                 );
               }),
         ),